refactor: cleanup packages
