+++ tests/test_evaluate.py
@@ -5,7 +5,7 @@ from unittest.mock import patch, MagicMock
 
 sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
 
-from wafsmith.cmd.evaluate import evaluate, EvaluateConfig, TestingEnv
+from wafsmith.cmd.evaluate import evaluate, EvaluateConfig
 
 
 @pytest.fixture
